fix/java: fix JNI bindings workign on desktop
1. We manually reimplemented the disconnect notifier callback
because bindgen has no knowledge about the lifetimes of the
callback objects, assuming that all of them are one-off calls.
Disconnect notifier can be called continuously, so to account
for this case we use a bindgen filter to filter out functions
that have the disconnect cb.

2. Desktop JNI implementation doesn't need the tricky class
loader caching as for Android. For desktop we just redirect
the find_class calls to use the native JNI library class loading
facilities.

3. ffi_utils and safe_bindgen dependencies have been updated to
include the fixes from the new versions.