chore/ci: provide Dockerfile for build
The Dockerfile here is quite simple. It runs the 'build-real' script to
cache all the dependencies. The dependencies for the non-mocked version
of SCL are a superset of the mocked dependencies, so we only need to
run the `build-real` script during the container build process. It also
runs the `package.rs` script with the --help option to cache the
dependencies for that script; downloading and compiling those can add a
couple of minutes into a build process.

A utility Makefile is provided for building the container. This will be
extended to provide some more useful targets for container management.

The build scripts have also been modified to accept a target directory
argument so that the same scripts can also be used in the container
build process. There's also a slight modification to remove linking each
line with `&&`, which I presume was done to stop the script on any
errors. The same can be achieved by using `set -e`, which in my opinion
makes the script a little less verbose.