chore/ci: pipeline for SCL build in Jenkins
This script was originally developed in the 'safe-build-infrastructure'
repo then moved into this one. It wasn't decided at that point if we
would keep the Jenkinsfile alongside the code or if it would perhaps go
into another repo.

The most important thing to note at the moment is that this code assumes
that the entire build runs end to end on a single slave. This
constraint is imposed by the use of a `docker commit` to make it more
convenient to work with the container. This may need refactored later.