chore/ci: introduce utility makefile
This is primarily intended to wrap the use of Docker for developers, for
making this mechanism easier to use. It handles awkward things like
mounting in the current directory with the correct permissions.

The reason why there's only a target for building safe_app and not the
others is because the build scripts don't currently have a separate
script for building safe_authenticator. The `build-real` script builds
both safe_authenticator and safe_app. There's a script for building core
on its own, but it didn't seem to make sense to add a target for core
and not one for authenticator, so I've just went with safe_app for now.