chore/ci: change 'warmup' stage to 'build'
I think name 'build' better captures the intent of this stage now, since
we will use the artifacts produced in this stage during the testing,
packaging, and deployment stages. We should save time by not having to
rebuild anything later, as the packaging currently does now.

This runs mocked and non-mocked builds for both Linux and OSX, all of
which run in parallel. Each job in a stage runs on a fresh VM, so we
need to make the artifacts we've built available to other stages, and
this is the reason for the deploy block for each stage. Unfortunately
the expression of this is quite verbose, but it seems unavoidable.

For Linux, Docker is introduced as the build mechanism. We just make use
of the Makefile that already exists, since it wraps the tedious detail
for using the container. It should be quite obvious that we're referring
to a Makefile at the root of this repository, so if anyone wants to know
what we're doing, they can just consult the Makefile.

Note that the S3 bucket names and branch conditionals used here will be
replaced after development.