refactor/ci: use Travis cache rather than S3
It turned out to be difficult to do the pull from S3. Even though the
artifacts that are uploaded are set to be publicly accessible, you can't
pull a directory from S3 without being authenticated. This in itself
isn't a problem. The issue was with providing the credentials. You can
define them using encrypted environment variables, but the problem is
those don't get made available to forks.

From the documentation:
"Encrypted environment variables are not available to pull requests
from forks due to the security risk of exposing such information to
unknown code."
https://docs.travis-ci.com/user/environment-variables/#defining-encrypted-variables-in-travisyml

We could of course zip the artifacts and upload them, but using the
Travis cache doesn't seem to be a bad solution, and it makes the Travis
file significantly less verbose.