chore/ci: run linux mock tests in container
After pulling the previous stage artifacts from the Travis cache, run
the tests against them inside the container. The artifacts get copied to
the root of the repo, which is then shared inside the container. There's
a little wrapper script that puts the artifacts in the correct location
before the tests run. Please see the comments on the script for the
reasoning behind that.