chore/ci: run compat tests in container
Runs the binary compatibility tests in the container. The test wrapper
is making an assumption that the previous set of tests will be mounted
in at /bct/tests, which I think is reasonable. The error handling tries
to provide as much information as possible for what's required to
execute the tests successfully.

The `after_script` command will overwrite the previous tests and make
this new set available to the next build.