refactor/ci: better artifact management
This creates the artifacts directory in the Makefile, since you also
want this to occur when you're using the Makefile locally.

The Travis cache location for storing the artifacts is also created
after the build has run successfully, since it makes sense that this
would be done after a successful build.