chore/ci: new packaging and deploy process
Changes the deployment process to take place on a single node.

The process is roughly as follows:
* Extract all previously built artifacts from the Travis cache to an
'artifacts' folder
* Run the container with this artifacts folder available in the mounted
volume at the app source directory
* Run a script in the container that wraps packaging for mocked and
non-mocked versions of safe_app and safe_authenticator for each OS
* The zip files to deploy will be generated in a 'deploy' directory that
will be made available outside the container because it is created in
the same directory as the shared volume

There's a slight modification to the packaging script to allow the
caller to pass in a location to look for artifacts. It will default to
CARGO_TARGET_DIR or simply 'target' if that's not provided.