refactor/ci: use S3 rather than Travis cache
It turned out that the Travis cache was unreliable for artifact
management. The docs on the cache actually say you shouldn't use it for
"compiled binaries".

Now the artifacts from the build stage are uploaded and retrieved from
S3. We need to upload them as an archive rather than using a folder on
S3, because even if the folder is public, you can't list a directory
without being authenticated. The problem with that is the keys would
need to be shared as encrypted environment variables, but the
documentation says these won't be made available to forked pull
requests, which makes it too inconvenient for our use.

Unfortunately this does make the Travis file a lot more verbose, but
there's not much way to get around this. At least not that I can see.