refactor/ci: artifact retrieval to Makefile
The artifact retrieval is placed in Makefile targets. This is so that
Jenkins will be able to use the exact same mechanism.

Makefiles are not so good for dealing with parameters. Really you need
to use environment variables, and while I do accept this is clunky, it
shouldn't be a huge problem given the vast majority of the time these
targets will be called in the build environment, which should have them
defined correctly.

You can see that all the stages declare the necessary environment
variables.