refactor/ci: do strip at build stage
The package.rs script attempts to run GNU strip on the binaries you want
to package. Strip needs to be run on the same OS the binary is built on,
so it is now performed at the build stage.

A flag is added to package.rs so that you need to tell it if you want to
have the strip performed.