refactor/build: new deployment process
The packaging stage is removed and this is now done during the
deployment stage. All the build artifacts are retrieved from S3 and
extracted into a little directory structure that the packaging process
uses to zip the artifacts to be deployed. A --artifacts argument is
added to package.rs to point it to the artifacts to be packaged.
Everything ends up in a deploy directory, which is then uploaded to S3.
This all takes place in the container, so there's a little wrapper
script added to coordinate everything.

In the Travis build, GNU strip runs during the packaging process. This
has been changed to be performed when the artifacts are produced, since
strip needs to run on the target OS. There was a conditional added to
package.rs to control this, so it wouldn't fail during the Jenkins
build. The --strip argument has been added to the Travis and Appveyor
scripts.