chore/ci: package artifacts as zip and tar.gz
Extends the packaging process to produce a tar.gz archive in addition to
a zip archive. I haven't created any flags or arguments to control what
types of archives are produced, assuming it's OK to just produce both.

This feature was requested by one of the front end products because the
build tooling they're using needs to work with a tar rather than a zip.