refactor/build: define params in jenkinsfile
For a multibranch pipeline job, you can't define parameters on the job
itself using the job DSL plugin, so they need to be defined here.

It doesn't look like there's actually much advantage of declaring them
this way as opposed to just using normal variables for them, but it
could be they need to be set this way for chaining builds in the future.