chore/ci: use project specific slave
We had been using the generic slave for this project, but we need to
switch it over to a project specific slave. This will then be rebuilt
with Rust 1.37.

We also remove the container being versioned along with the application.
We don't really need that for a build container and we discovered from
another project that this system doesn't work anyway. When you try to
upgrade to a new version there is no container for that version, so it
doesn't work with the version change commit system. We may change the
tag later on from 'build' to perhaps the name of the branch, but this
works for now.